+++ persistent/test/TemplateTestImports.hs
@@ -16,6 +16,9 @@ import Database.Persist.TH as X
 import Test.Hspec as X
 import Data.Proxy as X
 import Data.Text as X (Text)
+import Data.Maybe
+import Control.Monad
+import Language.Haskell.TH.Syntax
 
 data Foo = Bar | Baz
     deriving (Show, Eq)